+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.22.0] - 2020-12-08
+
+This change introduces a bottom navigation component & drawer.
+
+### Added
+
+- SimpleBottomNavigation
+- SwipeableTemporaryDrawer
+  - this component is an Avatar that opens a drawer
+  - this reads the avatar_url from apollo cache
+
+### Changed
+
+- layout2 uses SimpleBottomNavigation
+- `notistack` snackbars positioned at top, for XS screens
+
 ## [v0.21.2] - 2020-12-08
 
 ### Added
@@ -402,6 +418,7 @@ New:
 
 ### Added
 
+[v0.22.0]: https://github.com/thiskevinwang/coffee-code-climb/compare/v0.21.2...v0.22.0
 [v0.21.2]: https://github.com/thiskevinwang/coffee-code-climb/compare/v0.21.1...v0.21.2
 [v0.21.1]: https://github.com/thiskevinwang/coffee-code-climb/compare/v0.21.0...v0.21.1
 [v0.21.0]: https://github.com/thiskevinwang/coffee-code-climb/compare/v0.20.3...v0.21.0